fix notification bug

+++ controller/LibraryMapController.js
@@ -10,7 +10,7 @@ exports.getLibraryMaps = (req, res) => {
     });
 }
 
-exports.updateLibraryMap = (req, res) => {
+exports.createLibraryMap = (req, res) => {
     const file = req.file;
     const floor_name = req.body.floorName;
     const name = req.body.name;
@@ -20,14 +20,30 @@ exports.updateLibraryMap = (req, res) => {
         name,
     }
 
+    console.log(entry);
+
     LibraryMapRepository.findOrCreateLibraryMap(entry).spread(function (libraryMap, created) {
-        try {
-            fs.unlinkSync(libraryMap.image_url);
-        } catch (err) {
-            console.log(err.toString());
+        if (file !== undefined) {
+            try {
+                fs.unlinkSync(libraryMap.image_url);
+            } catch (err) {
+                console.log(err.toString());
+            }
+            libraryMap.image_url = file.path.replace(/\\/g, "/");
         }
-        libraryMap.image_url = file.path.replace(/\\/g, "/");
         libraryMap.save();
+        res.json({ success: true });
+    }).catch(err => {
+        res.status(500).json({ error: err.toString() });
+    });
+}
+
+exports.updateLibraryMapDetail = (req, res) => {
+    const id = req.body.id;
+    const floor_name = req.body.floorName;
+    const name = req.body.name;
+
+    LibraryMapRepository.updateLibraryMapDetail(id, floor_name, name).then((libraryMap) => {
         res.json({ success: true });
     }).catch(err => {
         res.status(500).json({ error: err.toString() });
